requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.16.0
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
blinker==1.5
boto3==1.26.81
botocore==1.29.81
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
dataclasses-json==0.5.7
decorator==5.1.1
deeplake==3.2.11
dill==0.3.6
entrypoints==0.4
filelock==3.9.0
frozenlist==1.3.3
git-ignore==1.0.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
hub==3.0.1
huggingface-hub==0.12.1
humbug==0.2.8
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.96
llama-index==0.4.15
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nltk==3.8.1
numcodecs==0.11.0
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.5.3
pathos==0.3.0
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
pox==0.3.2
ppft==1.7.6.6
protobuf==3.20.3
pyarrow==11.0.0
pycares==4.3.0
pycparser==2.21
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
streamlit==1.19.0
tenacity==8.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
yarl==1.8.2
zipp==3.15.0
